+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unversioned
 ### Fixed
 - \#3790 - Remove Docker on submit if no image is provided
+- \#3774 - External volume container tooltip needs update
 
 ## 1.1.2 - 2016-04-14
 ### Fixed