+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - \#3790 - Remove Docker on submit if no image is provided
 - \#3774 - External volume container tooltip needs update
 - \#3773 - Volume detail lacks "Volumes" in breadcrumb
+- \#3772 - Filter badge for External volumes is hidden until checked
 
 ## 1.1.2 - 2016-04-14
 ### Fixed