+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - \#3774 - External volume container tooltip needs update
 - \#3773 - Volume detail lacks "Volumes" in breadcrumb
 - \#3772 - Filter badge for External volumes is hidden until checked
+- \#3771 - Volume tab should not be showing in task detail view
 
 ## 1.1.2 - 2016-04-14
 ### Fixed