+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - \#3773 - Volume detail lacks "Volumes" in breadcrumb
 - \#3772 - Filter badge for External volumes is hidden until checked
 - \#3771 - Volume tab should not be showing in task detail view
+- \#3770 - External Volumes validation
 
 ## 1.1.2 - 2016-04-14
 ### Fixed